refactor : 패키지 이름 요구사항에 맞게 변경
