+++ src/main/java/domain/BonusNumber.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.List;
