+++ src/main/java/domain/WinningResult.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
