+++ src/main/java/domain/WinningStatus.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.Arrays;
 import java.util.Comparator;
