+++ src/test/java/domain/BonusNumberTest.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
