+++ src/test/java/domain/LottoTest.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.Arrays;
 import org.assertj.core.api.Assertions;
