+++ src/test/java/util/LottoGeneratorTest.java
@@ -3,7 +3,7 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.HashSet;
 import java.util.Set;
-import model.Lotto;
+import domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
